+++ app/src/main/java/com/thequietz/travelog/record/viewmodel/RecordBasicViewModel.kt
@@ -12,6 +12,9 @@ import com.thequietz.travelog.record.model.RecordBasicItem
 import com.thequietz.travelog.record.model.RecordImage
 import com.thequietz.travelog.record.repository.RecordBasicRepository
 import com.thequietz.travelog.schedule.model.ScheduleDetailModel
+import com.thequietz.travelog.util.createDateFromDay
+import com.thequietz.travelog.util.createDayFromDate
+import com.thequietz.travelog.util.nextDate
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -141,76 +144,6 @@ class RecordBasicViewModel @Inject constructor(
         return true
     }
 
-    // TODO: DateUtil와 같은 클래스로 분리 예정
-    private fun createDayFromDate(startDate: String, date: String): String {
-        val tempStartDate = startDate.split('.').map { it.toInt() }
-        val tempDate = date.split('.').map { it.toInt() }
-        val isLeapYear = tempDate[0] % 4 == 0
-        val dayOfMonth =
-            if (isLeapYear) listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-            else listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-
-        // 2020.12.30 ~ 2021.1.4
-        var day = 0
-        if (tempStartDate[0] < tempDate[0]) {
-            day = dayOfMonth[tempStartDate[1]] - tempStartDate[2]
-
-            for (i in tempStartDate[1] + 1..12) {
-                day += dayOfMonth[i]
-            }
-
-            day += tempDate[2]
-
-            for (i in tempDate[1] - 1 downTo 1) {
-                day += dayOfMonth[i]
-            }
-        }
-        // 2021.1.4 ~ 2021.3.5
-        else if (tempStartDate[1] < tempDate[1]) {
-            day = dayOfMonth[tempStartDate[1]] - tempStartDate[2]
-
-            for (i in tempStartDate[1] + 1 until tempDate[1]) {
-                day += dayOfMonth[i]
-            }
-
-            day += tempDate[2]
-        }
-        // 2021.1.4 ~ 2021.1.5
-        else {
-            day = tempDate[2] - tempStartDate[2]
-        }
-
-        return "Day${day + 1}"
-    }
-
-    private fun createDateFromDay(startDate: String, day: String): String {
-        val tempDate = startDate.split('.').map { it.toInt() }
-        val tempDay = day.substring(3).toInt() - 1
-        val isLeapYear = tempDate[0] % 4 == 0
-        val dayOfMonth =
-            if (isLeapYear) listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-            else listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-
-        // TODO("2달 이상 고려해야함.")
-
-        if (tempDate[2] + tempDay <= dayOfMonth[tempDate[1]]) {
-            return "${tempDate[0]}.${tempDate[1].toStringDate()}.${(tempDate[2] + tempDay).toStringDate()}"
-        }
-
-        if (tempDate[1] + 1 <= 12) {
-            return "${tempDate[0]}.${(tempDate[1] + 1).toStringDate()}.${((tempDate[2] + tempDay) % dayOfMonth[tempDate[1]]).toStringDate()}"
-        }
-
-        return "${tempDate[0] + 1}.01.${((tempDate[2] + tempDay) % dayOfMonth[tempDate[1]]).toStringDate()}"
-    }
-
-    private fun Int.toStringDate(): String {
-        if (this / 10 == 0) {
-            return "0$this"
-        }
-        return "$this"
-    }
-
     private fun createRecordBasicFromRecordImages(recordImages: List<RecordImage>): RecordBasic {
         val recordDestinationList = mutableListOf<RecordBasicItem.TravelDestination>()
 
@@ -263,24 +196,6 @@ class RecordBasicViewModel @Inject constructor(
         }
 
         return list.toList()
-    }
-
-    private fun String.nextDate(): String {
-        val tempDate = this.split('.').map { it.toInt() }
-        val isLeapYear = tempDate[0] % 4 == 0
-        val dayOfMonth =
-            if (isLeapYear) listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-            else listOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-
-        if (tempDate[2] + 1 <= dayOfMonth[tempDate[1]]) {
-            return "${tempDate[0]}.${(tempDate[1]).toStringDate()}.${(tempDate[2] + 1).toStringDate()}"
-        }
-
-        if (tempDate[1] + 1 <= 12) {
-            return "${tempDate[0]}.${(tempDate[1] + 1).toStringDate()}.01"
-        }
-
-        return "${tempDate[0] + 1}.01.01"
     }
 
     /* 이미지 추가 기능 삭제 예정
